update scripts for the stability measurements
